+++ src/main/java/frc/robot/vision/Align.java
@@ -1,0 +1,5 @@
+package frc.robot.vision;
+
+public class Align {
+
+}
