+++ src/main/java/frc/robot/vision/CMode.java
@@ -1,0 +1,7 @@
+package frc.robot.vision;
+
+// Different Camera Modes for setCameraMode method
+public enum CMode{
+    Drive,
+    Vision;
+}
